requirements.txt
numpy
pandas
Pillow
minisom